requirements.txt
# Core dependencies
pyjwt>=2.0.0
pandas>=1.0.0
requests>=2.25.0
websockets>=10.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
